+++ Tests/BugsnagThreadTests.m
@@ -1,5 +1,5 @@
 //
-//  BugsnagThreadTest.m
+//  BugsnagThreadTests.m
 //  Tests
 //
 //  Created by Jamie Lynch on 07/04/2020.
@@ -8,15 +8,22 @@
 
 #import <XCTest/XCTest.h>
 
+#import "BSG_KSMachHeaders.h"
 #import "BugsnagStackframe.h"
 #import "BugsnagThread+Private.h"
+#import "BugsnagThread+Recording.h"
 
-@interface BugsnagThreadTest : XCTestCase
+@interface BugsnagThreadTests : XCTestCase
 @property NSArray *binaryImages;
 @property NSDictionary *thread;
 @end
 
-@implementation BugsnagThreadTest
+@implementation BugsnagThreadTests
+
++ (void)setUp {
+    bsg_mach_headers_initialize();
+    bsg_mach_headers_register_for_changes();
+}
 
 - (void)setUp {
     self.thread = @{
@@ -201,6 +208,63 @@
     XCTAssertEqual(1, thread.stacktrace.count);
     thread.stacktrace = @[];
     XCTAssertEqual(0, thread.stacktrace.count);
+}
+
+// MARK: - BugsnagThread (Recording)
+
+- (void)testAllThreads {
+    NSArray<BugsnagThread *> *threads = [BugsnagThread allThreads:YES callStackReturnAddresses:NSThread.callStackReturnAddresses];
+    XCTAssertTrue(threads[0].errorReportingThread);
+    XCTAssertEqualObjects(threads[0].name, @"com.apple.main-thread");
+    XCTAssertEqualObjects(threads[0].stacktrace.firstObject.method, @(__PRETTY_FUNCTION__));
+    XCTAssertGreaterThan(threads.count, 1);
+}
+
+- (void)testCurrentThread {
+    NSArray<BugsnagThread *> *threads = [BugsnagThread allThreads:NO callStackReturnAddresses:NSThread.callStackReturnAddresses];
+    XCTAssertEqual(threads.count, 1);
+    XCTAssertTrue(threads[0].errorReportingThread);
+    XCTAssertEqualObjects(threads[0].id, @"0");
+    XCTAssertEqualObjects(threads[0].name, @"com.apple.main-thread");
+    XCTAssertEqualObjects(threads[0].stacktrace.firstObject.method, @(__PRETTY_FUNCTION__));
+}
+
+- (void)testCurrentThreadFromBackground {
+    __block BugsnagThread *thread = nil;
+    XCTestExpectation *expectation = [self expectationWithDescription:@"Thread recorded in background"];
+    dispatch_async(dispatch_get_global_queue(QOS_CLASS_USER_INITIATED, 0), ^{
+        thread = [BugsnagThread allThreads:NO callStackReturnAddresses:NSThread.callStackReturnAddresses][0];
+        [expectation fulfill];
+    });
+    [self waitForExpectationsWithTimeout:2 handler:nil];
+    XCTAssertTrue(thread.errorReportingThread);
+    XCTAssertGreaterThan(thread.id.intValue, 0);
+    XCTAssertNotNil(thread.name);
+    XCTAssertNotEqualObjects(thread.name, @"com.apple.main-thread");
+    XCTAssert([thread.stacktrace.firstObject.method hasSuffix:@"_block_invoke"]);
+}
+
+- (void)testMainThread {
+    BugsnagThread *thread = [BugsnagThread mainThread];
+    XCTAssertNil(thread);
+}
+
+#define XCTAssertContainsObject(collection, object) \
+    XCTAssert([collection containsObject:object], @"%@ is not contained in %@", object, \
+        [[collection description] stringByReplacingOccurrencesOfString:@"\n" withString:@"\\n"])
+
+- (void)testMainThreadFromBackground {
+    __block BugsnagThread *thread = nil;
+    dispatch_semaphore_t semaphore = dispatch_semaphore_create(0);
+    dispatch_async(dispatch_get_global_queue(QOS_CLASS_USER_INITIATED, 0), ^{
+        thread = [BugsnagThread mainThread];
+        dispatch_semaphore_signal(semaphore);
+    });
+    dispatch_semaphore_wait(semaphore, dispatch_time(DISPATCH_TIME_NOW, 2 * NSEC_PER_SEC));
+    XCTAssertTrue(thread.errorReportingThread);
+    XCTAssertEqualObjects(thread.id, @"0");
+    XCTAssert([thread.name isEqualToString:@"com.apple.main-thread"] || [thread.name isEqualToString:@"com.apple.dt.xctest.xctwaiter"]);
+    XCTAssertContainsObject([thread.stacktrace valueForKeyPath:@"method"], @(__PRETTY_FUNCTION__));
 }
 
 @end