+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+## TBD
+
+### Bug fixes
+
+* Improve performance of thread recording for handled errors.
+  [#992](https://github.com/bugsnag/bugsnag-cocoa/pull/992)
+
 ## 6.6.2 (2021-02-10)
 
 ### Bug fixes